+++ frontend/twitter-clone/src/api/authApi.ts
@@ -2,6 +2,8 @@ import { LoginUser, User } from "../types/User";
 import { axiosClient } from "./axiosClient";
 
 const authApi = {
+	// メールチェックAPI
+	existEmail: (email: string) => axiosClient.post("auth/authUser", { email }),
 	// 新規登録API
 	register: (params: User) => axiosClient.post("auth/register", params),
 	// ログインAPI
@@ -9,15 +11,14 @@ const authApi = {
 	// JWT認証API
 	verifyToken: (token: string) =>
 		axiosClient.post("auth/verify-token", { token }),
-	// パスワード忘れリクエストAPI
+	// パスワードリセット通知API
 	forgotPassword: (param: {
 		username: string | undefined;
 		email: string | undefined;
 	}) => axiosClient.post("auth/forgotPassword", param),
 	// パスワードリセットAPI
 	resetPassword: (param: {
-		username: string | undefined;
-		email: string | undefined;
+		resetToken: string;
 		password: string;
 		confirmPassword: string;
 	}) => axiosClient.post("auth/resetPassword", param),
